# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Games-API
  version: 1.0.0
  description: '###This is a Swagger Ui for our Games Api###'
paths:
  /games:
    description: ''
    get:
      tags:
        - games
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: {}
              examples:
                List of games:
                  value:
                    - Witcher 3
                    - Cyberpunk 2077
      summary: Get a list of all games
      description: Returns a list of all games in API
    post:
      tags:
        - new-game
      summary: Creates a game
      description: Creates a new game
  '/games/{id}':
    get:
      tags:
        - details
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                Full details of a game:
                  value: '    {id: 1, name: "Witcher 3", price: 29.99},'
          description: Returns full details of a given id game
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Game /w+ was not found:
                  value:
                    error: Game \w+ was not found in game array
          description: Game was not found in games array
      operationId: id
      summary: Get full details of a game
      description: Returns full details of a given id game
    post:
      requestBody:
        description: A new game to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/error'
            examples:
              Entity id was not given:
                value:
                  error: Entity id is not given
        required: true
      tags:
        - new-game
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/new-game'
          description: Returns created game details
        '400':
          description: One or all params are missing
      summary: Creates a game
      description: Creates a new game
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/delete-game'
      tags:
        - delete-game
      responses:
        '204':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/delete-game'
          description: There is no content to return
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Game /w+ was not found:
                  value:
                    error: Game /w+ was not found to delete
          description: Game was not found in games array
      summary: Deletes game
      description: Deletes an existing game
components:
  schemas:
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
    new-game:
      title: Root Type for new-game
      description: ''
      required:
        - price
        - name
      type: object
      properties:
        name:
          type: string
        price:
          format: float
          type: number
      example:
        name: Tetris
        price: 0.99
    delete-game:
      title: Root Type for delete-game
      description: ''
      required:
        - id
      type: object
      properties:
        id:
          format: int32
          type: integer
      example:
        id: 1